+++ codemods/misc/migrate-mui-tooltip-to-bui-tooltip/tests/placement-dropped/expected.tsx
@@ -1,9 +1,11 @@
 import { Tooltip, TooltipTrigger } from '@backstage/ui';
 
 const MyComponent = () => (
+  <>
   {/* TODO(backstage-codemod): verify Tooltip placement mapping manually */}
-<TooltipTrigger>
+  <TooltipTrigger>
   <span>Hover</span>
   <Tooltip>Info</Tooltip>
 </TooltipTrigger>
+</>
 );