+++ codemods/misc/migrate-mui-typography-to-text/tests/noop-no-mui-import/input.tsx
@@ -1,5 +1,0 @@
-import { Text } from '@backstage/ui';
-
-const MyComponent = () => (
-  <Text variant="body">Already migrated</Text>
-);