+++ codemods/misc/migrate-mui-typography-to-text/CHANGELOG.md
@@ -1,1 +1,7 @@
 # @backstage/migrate-mui-typography-to-text
+
+## 0.1.0
+
+### Minor Changes
+
+- Initial release: migrate MUI `Typography` / `DialogContentText` to Backstage UI `Text` during the MUI 4 to BUI migration.